attempt to fix bugs

+++ Raspberry Pi/config/config.py
@@ -6,7 +6,7 @@ RES = [640,480]
 # Number of slices in the Image:
 NUMofCUT = 10
 # Gives an array of the slices' x coordinate.
-CUTHEIGHT = np.linspace(RES[1]-10,20,num = NUMofCUT).astype(np.uint8)
+CUTHEIGHT = np.linspace(RES[1]-10,20,num = NUMofCUT).astype(int)
 # Percentage of tolerance when detecting unexpected points:
 tol = 0.3 * RES[0]
 CENTER = int(RES[0]/2)
